+++ src/examples/experimental/wscript
@@ -3,9 +3,15 @@
 
 
 def build(bld):
-  obj = bld.program(features = 'cxx',
-                         source = 'parego.cpp',
-                         includes = '.. ../.. ../../../',
-                         target = 'parego',
-                         uselib = 'BOOST EIGEN TBB SFERES LIBCMAES NLOPT',
-                         use = 'limbo')
+    obj = bld.program(features='cxx',
+                      source='imgpo.cpp',
+                      includes='. .. ../../',
+                      target='imgpo',
+                      uselib='BOOST EIGEN TBB LIBCMAES NLOPT',
+                      use='limbo')
+  # obj = bld.program(features = 'cxx',
+  #                        source = 'parego.cpp',
+  #                        includes = '.. ../.. ../../../',
+  #                        target = 'parego',
+  #                        uselib = 'BOOST EIGEN TBB SFERES LIBCMAES NLOPT',
+  #                        use = 'limbo')
